requirements.txt
faiss-cpu==1.12.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
torch==2.0.1
transformers==4.30.2
fastapi==0.100.0
uvicorn==0.22.6
pydantic==2.0.3
pytest==7.4.0
python-dotenv==1.0.0
